api/pyproject.toml
[project]
name = "django-atomicserver-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=4.2,<6.0",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.0",
]
